Reduce circular imports by avoiding to import Main from everywhere. Many modules really only need access to the parse Context object, which can be extracted into a separate module.
