+++ Cython/Tests/TestJediTyper.py
@@ -11,7 +11,7 @@ from contextlib import contextmanager
 from tempfile import NamedTemporaryFile
 
 from Cython.Compiler.ParseTreeTransforms import NormalizeTree, InterpretCompilerDirectives
-from Cython.Compiler import Main, Symtab, Visitor, Options
+from Cython.Compiler import ParseContext, Symtab, Visitor, Options
 from Cython.TestUtils import TransformTest
 
 TOOLS_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', 'Tools'))
@@ -211,7 +211,7 @@ class TestTypeInjection(TestJediTyper):
     def setUp(self):
         super(TestTypeInjection, self).setUp()
         compilation_options = Options.CompilationOptions(Options.default_options)
-        ctx = Main.Context.from_options(compilation_options)
+        ctx = ParseContext.Context.from_options(compilation_options)
         transform = InterpretCompilerDirectives(ctx, ctx.compiler_directives)
         transform.module_scope = Symtab.ModuleScope('__main__', None, ctx)
         self.declarations_finder = DeclarationsFinder()
